+++ crates/ludtwig/src/main.rs
@@ -1,5 +1,8 @@
 #![warn(clippy::pedantic)]
 #![allow(clippy::module_name_repetitions)]
+#![allow(clippy::doc_markdown)]
+#![allow(clippy::elidable_lifetime_names)]
+#![allow(clippy::needless_continue)]
 
 use crate::check::rule::{Rule, Severity};
 use crate::check::rules::get_config_active_rule_definitions;